+++ frontend/react-app/src/Header.jsx
@@ -1,8 +1,7 @@
 import { Box, Breadcrumbs, makeStyles, Typography } from '@material-ui/core'
 import { deepPurple } from '@material-ui/core/colors';
-import { useEffect } from 'react';
 import { Link, useLocation } from 'react-router-dom';
-import { routes } from './routes'
+import { useHeader } from './hooks'
 
 const useStyles = makeStyles(() => ({
   link: {
@@ -15,27 +14,22 @@ const useStyles = makeStyles(() => ({
 export const Header = () => {
   const classes = useStyles();
   const location = useLocation();
-  const path = location.pathname;
-  const route = routes.find(route => route.path === path)
-  const paths = path.split('/').filter(path => path)
-  const names = ['ホーム'].concat(route.name.split(' | ').filter(name => name));
+  const { names, paths, status } = useHeader(location);
 
-  useEffect(() => {
-    if (route.name) {
-      document.title = 'Algorithms Simulator | ' + route.name
+  return (<>
+    {status &&
+      <Breadcrumbs aria-label="breadcrumb" separator="›">
+        {['ホーム'].concat(paths).map((path, index) => <Box key={index}>
+          {index !== paths.length && <Link to={'/' + paths.slice(0, index).join('/')}><Typography color="textSecondary" className={classes.link}>{names[index]}</Typography></Link>}
+          {index === paths.length && <Typography color="textPrimary">{names[index]}</Typography>}
+        </Box>
+        )}
+      </Breadcrumbs>
     }
-    else {
-      document.title = 'Algorithms Simulator'
+    {!status &&
+      <Breadcrumbs aria-label="breadcrumb" separator="›">
+        <Typography color="textPrimary">{"ページが存在しません"}</Typography>
+      </Breadcrumbs>
     }
-  }, [route])
-
-  return (
-    <Breadcrumbs aria-label="breadcrumb" separator="›">
-      {['ホーム'].concat(paths).map((path, index) => <Box key={index}>
-        {index !== paths.length && <Link to={'/' + paths.slice(0, index).join('/')}><Typography color="textSecondary" className={classes.link}>{names[index]}</Typography></Link>}
-        {index === paths.length && <Typography color="textPrimary">{names[index]}</Typography>}
-      </Box>
-      )}
-    </Breadcrumbs>
-  )
+  </>)
 }
